+++ EditorEX/AudioSpectrogram/Managers/SpectrogramManager.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace EditorEX.AudioSpectrogram.Managers
+{
+    internal class SpectrogramManager
+    {
+    }
+}
