--- EditorEX/Utilities/SceneUtil.cs
@@ -1,13 +1,0 @@
-﻿using UnityEngine.SceneManagement;
-
-namespace EditorEX.Utilities
-{
-    internal static class SceneUtil
-    {
-        public static bool IsInBeatmapEditor()
-        {
-            var activeScene = SceneManager.GetActiveScene();
-            return activeScene.name == "BeatmapEditor3D" || activeScene.name == "BeatmapLevelEditorWorldUi";
-        }
-    }
-}
